requirements.txt
# Django
Django==2.0.1
django-environ==0.4.4

# APIStar
apistar==0.3.9

# Database
psycopg2==2.7.3.2

# Tests
pytest==3.4.0
pytest-django==3.1.2
